Fix prevent user to set old password on password-reset

+++ backend/app/Services/Application/Handlers/Auth/ResetPasswordHandler.php
@@ -3,6 +3,7 @@
 namespace HiEvents\Services\Application\Handlers\Auth;
 
 use HiEvents\DomainObjects\UserDomainObject;
+use HiEvents\Exceptions\PasswordInvalidException;
 use HiEvents\Mail\User\ResetPasswordSuccess;
 use HiEvents\Repository\Interfaces\PasswordResetTokenRepositoryInterface;
 use HiEvents\Repository\Interfaces\UserRepositoryInterface;
@@ -38,11 +39,20 @@ class ResetPasswordHandler
             $resetToken = $this->passwordTokenValidateService->validateAndFetchToken($resetPasswordData->token);
             $user = $this->validateUser($resetToken->getEmail());
 
+            if ($this->checkNewPasswordIsOldPassword($user, $resetPasswordData->password)) {
+                throw new PasswordInvalidException(__('New password must be different from the old password.'));
+            }
+
             $this->resetUserPassword($user->getId(), $resetPasswordData->password);
             $this->deleteResetToken($resetToken->getEmail());
             $this->logResetPasswordSuccess($user);
             $this->sendResetPasswordEmail($user);
         });
+    }
+
+    private function checkNewPasswordIsOldPassword(UserDomainObject $user, string $newPassword): bool
+    {
+        return $this->hashManager->check($newPassword, $user->getPassword());
     }
 
     private function validateUser(string $email): UserDomainObject
@@ -72,7 +82,7 @@ class ResetPasswordHandler
         $this->passwordResetTokenRepository->deleteWhere(['email' => $email]);
     }
 
-    private function logResetPasswordSuccess($user): void
+    private function logResetPasswordSuccess(UserDomainObject $user): void
     {
         $this->logger->info('Password reset successfully', [
                 'user_id' => $user->getId(),
